Inventory Display implemented.
Inventory Display is a grid template with 5 items per row. The specifics for the grid itself is in the css file with the class .inventory_grid.

The cards themselves are bootstrap cards, with the headers containing a star that displays the rarity, the pet image, and the pet rarity and species at the card body. To be implemented here is modal function for the pet view.

Next to be implemented are the overflow and scroll, before moving onto the start of the filters.