Additional Inventory features implemented.
Added pagination on the Inventory display that splices the pet objects into different pages. The pages can be accessed through links at the bottom of the display indicating "Previous" and "Next". Links are also promptly disabled if there are no other pages before or after the current page.

Side note, The pet cards are now linkable, which will be useful in targeting modals and IDs for the pet view. The cards also change colors when hovered over, indicating their rarity.

Next to be added will be filterization and forms relating to such, which will either sort or filter the current object list. After that will be the pet display view.